+++ api/models/v0.3/modelPolicy.js
@@ -53,11 +53,16 @@ const Policy = (args) => {
         },
       },
       cardType: {
-        standard: IlsClient.WEB_APPLICANT_EXPIRATION_TIME,
+        standard: IlsClient.STANDARD_EXPIRATION_TIME,
         temporary: IlsClient.WEB_APPLICANT_EXPIRATION_TIME,
       },
-      requiredFields: ["birthdate"],
+      requiredFields: ["email", "barcode", "birthdate"],
       minimumAge: 13,
+      serviceArea: {
+        city: ALLOWED_CITIES,
+        county: ALLOWED_COUNTIES,
+        state: ALLOWED_STATES,
+      },
     },
     simplyeJuvenile: {
       agency: IlsClient.DEFAULT_PATRON_AGENCY,
@@ -69,9 +74,9 @@ const Policy = (args) => {
       },
       cardType: {
         standard: IlsClient.STANDARD_EXPIRATION_TIME,
-        temporary: IlsClient.TEMPORARY_EXPIRATION_TIME,
+        temporary: IlsClient.STANDARD_EXPIRATION_TIME,
       },
-      requiredFields: ["barcode"],
+      requiredFields: ["email", "barcode"],
       serviceArea: {
         city: ALLOWED_CITIES,
         county: ALLOWED_COUNTIES,