+++ app.js
@@ -1,32 +1,55 @@
-const express = require('express');
-const bodyParser = require('body-parser');
+const express = require("express");
+const bodyParser = require("body-parser");
 
 // Import controllers
-const createPatronV0_1 = require('./api/controllers/v0.1/createPatron.js'); // eslint-disable-line camelcase
-const validations = require('./api/controllers/v0.1/validations.js');
-const createPatronV0_2 = require('./api/controllers/v0.2/createPatron.js'); // eslint-disable-line camelcase
-const apiDoc = require('./api/controllers/apiDoc.js');
-const createPatronV0_3 = require('./api/controllers/v0.3/endpoints.js'); // eslint-disable-line camelcase
+const createPatronV0_1 = require("./api/controllers/v0.1/createPatron.js"); // eslint-disable-line camelcase
+const validations = require("./api/controllers/v0.1/validations.js");
+const createPatronV0_2 = require("./api/controllers/v0.2/createPatron.js"); // eslint-disable-line camelcase
+const apiDoc = require("./api/controllers/apiDoc.js");
+const createPatronV0_3 = require("./api/controllers/v0.3/endpoints.js"); // eslint-disable-line camelcase
+
+const awsDecrypt = require("./config/awsDecrypt");
 
 const app = express();
-const pathName = `${process.cwd()}/config/deploy_${app.get('env')}.env`;
-require('dotenv').config({ path: pathName });
+const pathName = `${process.cwd()}/config/deploy_${app.get("env")}.env`;
+require("dotenv").config({ path: pathName });
 
-const BarcodeDb = require('./db');
+const BarcodeDb = require("./db");
 
-// Initialize the connection to the database.
-// This is done here so we can have one instance and one connection to the
-// database for however many times the lambda is used for.
-const db = BarcodeDb({
-  user: process.env.DB_USER,
-  host: process.env.DB_HOST,
-  database: process.env.DB_DATABASE,
-  password: process.env.DB_PASSWORD,
-  port: process.env.DB_PORT,
-});
-// Once we get the database class instance, initialize it by creating the
-// table and inserting the seed data if it's not in the database already.
-db.init();
+// Decrypt any database credentials if on QA or production for the NYPL AWS
+// environment. Then create the single database instance in the app.
+async function initDatabase() {
+  let dbUser;
+  let dbHost;
+  let dbPassword;
+
+  if (process.env.NODE_ENV === "development") {
+    dbUser = process.env.DB_USER;
+    dbHost = process.env.DB_HOST;
+    dbPassword = process.env.DB_PASSWORD;
+  } else {
+    dbUser = await awsDecrypt.decryptKMS(process.env.DB_USER);
+    dbHost = await awsDecrypt.decryptKMS(process.env.DB_HOST);
+    dbPassword = await awsDecrypt.decryptKMS(process.env.DB_PASSWORD);
+  }
+
+  // Initialize the connection to the database.
+  // This is done here so we can have one instance and one connection to the
+  // database for however many times the lambda is used for.
+  const db = BarcodeDb({
+    user: dbUser,
+    host: dbHost,
+    password: dbPassword,
+    // These are constant and don't need to be encrypted or decrypted.
+    database: process.env.DB_DATABASE,
+    port: process.env.DB_PORT,
+  });
+  // Once we get the database class instance, initialize it by creating the
+  // table and inserting the seed data if it's not in the database already.
+  db.init();
+}
+
+initDatabase();
 
 // Below are the middlewares for response headers
 /**
@@ -38,13 +61,12 @@ db.init();
  * @param {next}
  */
 function allowCrossDomain(req, res, next) {
-  res.header('Access-Control-Allow-Origin', '*');
-  res.header('Access-Control-Allow-Methods', 'GET, PUT, POST, DELETE, OPTIONS');
+  res.header("Access-Control-Allow-Origin", "*");
+  res.header("Access-Control-Allow-Methods", "GET, PUT, POST, DELETE, OPTIONS");
   res.header(
-    'Access-Control-Allow-Headers',
-    'Origin, X-Requested-With, Content-Type, Accept',
+    "Access-Control-Allow-Headers",
+    "Origin, X-Requested-With, Content-Type, Accept",
   );
-
   next();
 }
 
@@ -74,9 +96,9 @@ function v1Error(err, req) {
     data: {
       status_code_from_card_creator: null,
       status: err.status,
-      type: 'invalid-request',
+      type: "invalid-request",
       message: `Request body: ${err.body}`,
-      detail: 'The patron creator did not forward the request to Card Creator.',
+      detail: "The patron creator did not forward the request to Card Creator.",
     },
     count: 0,
   };
@@ -94,9 +116,9 @@ function v2Error(err, req) {
   return {
     status_code_from_card_ils: null,
     status: err.status,
-    type: 'invalid-request',
+    type: "invalid-request",
     message: `Request body: ${err.body}`,
-    detail: 'The patron creator did not forward the request to the ILS.',
+    detail: "The patron creator did not forward the request to the ILS.",
   };
 }
 
@@ -115,17 +137,17 @@ function errorHandler(err, req, res, next) {
   // eslint-disable-next-line no-console
   console.error(
     `status_code: ${err.status}, `
-      + 'type: "invalid-request", '
+      + "type: \"invalid-request\", "
       + `message: "Request body: ${err.body}"` // eslint-disable-line comma-dangle
   );
   let jsonError;
 
-  if (req.url.includes('v0.1')) {
-    if (req.url.includes('validations')) {
+  if (req.url.includes("v0.1")) {
+    if (req.url.includes("validations")) {
       jsonError = validations.renderResponseData(
         null,
         false,
-        'invalid-request',
+        "invalid-request",
         null,
         `Error request with request body ${err.body}`,
         {},
@@ -133,7 +155,7 @@ function errorHandler(err, req, res, next) {
     } else {
       jsonError = v1Error(err, req);
     }
-  } else if (req.url.includes('v0.2')) {
+  } else if (req.url.includes("v0.2")) {
     jsonError = v2Error(err, req);
   }
 
@@ -147,26 +169,26 @@ app.use(errorHandler);
 const router = express.Router();
 
 // This route will make a request for swaggerDoc.json
-app.get('/docs/patrons-validations', apiDoc.renderApiDoc);
+app.get("/docs/patrons-validations", apiDoc.renderApiDoc);
 
-app.use('/api', router);
+app.use("/api", router);
 
-router.route('/v0.1/patrons/').post(createPatronV0_1.createPatron);
+router.route("/v0.1/patrons/").post(createPatronV0_1.createPatron);
 // Still supporting older v0.1 validation endpoints
-router.route('/v0.1/validations/username').post(validations.checkUserName);
-router.route('/v0.1/validations/address').post(validations.checkAddress);
+router.route("/v0.1/validations/username").post(validations.checkUserName);
+router.route("/v0.1/validations/address").post(validations.checkAddress);
 
 // New validation will be part of the `patrons` endpoint.
-router.route('/v0.2/patrons/').post(createPatronV0_2.createPatron);
+router.route("/v0.2/patrons/").post(createPatronV0_2.createPatron);
 
 // Still supporting older v0.1 validation endpoints
-router.route('/v0.3/validations/username').post(createPatronV0_3.checkUsername);
-router.route('/v0.3/validations/address').post(createPatronV0_3.checkAddress);
+router.route("/v0.3/validations/username").post(createPatronV0_3.checkUsername);
+router.route("/v0.3/validations/address").post(createPatronV0_3.checkAddress);
 router
-  .route('/v0.3/patrons/dependent-eligibility')
+  .route("/v0.3/patrons/dependent-eligibility")
   .get(createPatronV0_3.checkDependentEligibility);
-router.route('/v0.3/patrons/dependents').post(createPatronV0_3.createDependent);
-router.route('/v0.3/patrons/').post(createPatronV0_3.createPatron);
+router.route("/v0.3/patrons/dependents").post(createPatronV0_3.createDependent);
+router.route("/v0.3/patrons/").post(createPatronV0_3.createPatron);
 
 // Do not listen to connections in Lambda environment
 if (!process.env.AWS_LAMBDA_FUNCTION_NAME) {