+++ config/awsDecrypt.js
@@ -1,13 +1,13 @@
 // The credentials for NYPL's Simplified Card Creator API
-const aws = require('aws-sdk');
+const aws = require("aws-sdk");
 
 const kms = new aws.KMS({
-  region: 'us-east-1',
+  region: "us-east-1",
 });
 
 const decryptKMS = (key) => {
   const params = {
-    CiphertextBlob: Buffer.from(key, 'base64'),
+    CiphertextBlob: Buffer.from(key, "base64"),
   };
 
   return new Promise((resolve, reject) => {